requirements.txt
pika==1.3.2
boto3-1.28.54 
botocore-1.31.54 
jmespath-1.0.1 
s3transfer-0.6.2
pytest==7.4.2